1 more print fix

+++ src/platform/platform_posix.c
@@ -201,7 +201,7 @@ CxPlatUninitialize(
         uint32_t Tag = AllocatedTags.Tags[TagCount].Tag;
         printf("%c%c%c%c\t", Tag & 0xFF, (Tag >> 8) & 0xFF, (Tag >> 16) & 0xFF, (Tag >> 24) & 0xFF);
         for (uint32_t SizeCount = 0; SizeCount < AllocatedTags.Tags[TagCount].CurrentSizeCount; SizeCount++) {
-            printf("%lu(%lu)\t", AllocatedTags.Tags[TagCount].Sizes[SizeCount].Size, AllocatedTags.Tags[TagCount].Sizes[SizeCount].Count);
+            printf("%llu(%llu)\t", (unsigned long long)AllocatedTags.Tags[TagCount].Sizes[SizeCount].Size, (unsigned long long)AllocatedTags.Tags[TagCount].Sizes[SizeCount].Count);
         }
         printf("\n");
     }
